requirements.txt
requests
beautifulsoup4
pycryptodome
fake-useragent